Fixes problem when ext::next is called first
If ext::next is called without a call to ext::has_next before hand, then
previously it would return a default constructed matrix. Now a first call to
has_next is included in the constructor to ensure that the first call to next
gives a good matrix. This does not however fix the problem where next is called
repeatedly without any calls to has_next. I suspect that case is still broken.