Tried adding a fail fast elliptic check.
As each principle submatrix of a positive definite matrix is itself positive
definite, I added a check whether the vertex submatrix is positive definite
before adding any of the extra rows/columns. This should have allowed the code
to skip a number of the choleksy computations, but in reality ended up being
slower in the benchmarks for some reason.